Add Graphite export functionality
Graphite export is disabled by default. When setting the system property
'graphite.url' to the graphite database destination the export will be
enabled on startup.

By default the export happens every 15 seconds but it can be changed by
setting the system property 'graphite.interval' in seconds.

The metrics will appear in Graphite/Grafana with the prefix 'vdsmfake'.

Change-Id: Ic06804bc972ac5a5a71eb8689acbebb64669dae4
Signed-off-by: Roman Mohr <rmohr@redhat.com>