Use proper Config facility
Motivation:
Configuration are set in web.xml and runtime System properties and it is
hard to maintain that. We need a proper file based with overrides
support.

Modification:
Using a Config library(Typesafe Config) a config properties file is
loaded, and can be overriden by System Properties and can be reloadeble,
merge other sources etc (see typesafe config page)

Result:
The application uses a single AppConfig facade to get all the config
items. The order of configuration loading is (highest in order wins)
1. System Properties
2. ./vdsm.conf (missing by default)
3. application.conf (shipped inside vdsmfake.jar)

fixes: #5
Change-Id: Ife8135333bf8f3e37160e73ff1e406c537289229
Signed-off-by: Roy Golan <rgolan@redhat.com>