Remove the explicit shutdown of jsonRpcServer
Motivation:
There server class is exposing a static method a graceful shutdown,
using an internal static field which is modified by an instance  member.
It's better to just encapsulate the lifecycle management in the class
itself

Modification:
Make the inner state non-static
Remove static close() method
Register a shutdown-hook to close the JsonRpcServer

Result:
JsonRpcServer get called by a shutdown to clean it's resources
No need to call that externally.

Change-Id: I5bdbfaddc236d2a1136e9587418bfcce0567c3c1
Signed-off-by: Roy Golan <rgolan@redhat.com>