RPM Packaging
Added a build script to create an RPM.

Also the script is called from automation/build-artifacts.sh
to create artifacts on every merge + on demand jobs.

 ./build-rpm.sh

Final tweak to container creation - now there is a separate
'automation/Dockerfile' that consumes the generated RPM and installs it
rather then compiling everything from the start.
That makes the container smaller and faster to create.

Fixes: https://github.com/oVirt/ovirt-vdsmfake/issues/10
Change-Id: Id00fd5cf3c8a45c8cb2efec52be5186dc3500888
Signed-off-by: Roy Golan <rgolan@redhat.com>